requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
requests
chromadb
sentence-transformers
pypdf
markdown
unstructured
tiktoken
ollama